refactor: apis/ranking 폴더구조 변경
