+++ src/services/apis/ranking/ranking.type.ts
